Added examples, readme. Fixed some stuff

+++ src/UsbCallbacks.cpp
@@ -2,9 +2,9 @@
 #include <cstdint>
 #include <stddef.h>
 
-void UsbCallbacks::onRecievedMessage(const uint8_t *buffer, const size_t bufferLength) {}
+void UsbCallbacks::onRecievedMessage(const uint8_t *buffer, const uint16_t bufferLength) {}
 
-void UsbCallbacks::onSentMessage(const uint8_t *buffer, const size_t bufferLength) {}
+void UsbCallbacks::onSentMessage(const uint8_t *buffer, const uint16_t bufferLength) {}
 
 UsbCallbacks::UsbCallbacks() {}
