+++ src/UsbCallbacks.h
@@ -4,8 +4,8 @@
 class UsbCallbacks
 {
 public:
-  virtual void onRecievedMessage(const uint8_t *buffer, const size_t bufferLength);
-  virtual void onSentMessage(const uint8_t *buffer, const size_t bufferLength);
+  virtual void onRecievedMessage(const uint8_t *buffer, const uint16_t bufferLength);
+  virtual void onSentMessage(const uint8_t *buffer, const uint16_t bufferLength);
   UsbCallbacks();
   ~UsbCallbacks();
 };
